.github/actions/generate-global-acl/requirements.txt
gcsfs==2022.3.0
google-cloud-storage==2.3.0
click==8.1.3
universal_pathlib==0.0.18
PyYAML==6.0